Move tests to test_typing
